Move the view to api, #8282
